test-#145: write unit tests for REST customer dtos
