+++ src/microservices/customer/interfaces/update-customer-object.interface.ts
@@ -1,5 +1,5 @@
 import { UpdateCustomerProfileInput } from '../graphql/input/update-customer-profile';
-import { UpdateCustomerProfileDto } from '../rest/dto/update-customer-profile.dto';
+import { UpdateCustomerProfileDto } from '../rest/dto/update-customer-profile';
 
 export interface IUpdateCustomerObject {
   id: number;
