--- src/microservices/customer/rest/dto/create-customer-profile.dto.ts
@@ -1,5 +1,0 @@
-export class CreateCustomerProfileDto {
-  auth_id: number;
-  first_name: string;
-  last_name: string;
-}
