+++ src/microservices/customer/rest/dto/create-customer-profile/index.ts
@@ -1,0 +1,1 @@
+export * from './create-customer-profile.dto';
