--- src/microservices/customer/rest/dto/update-customer-profile.dto.ts
@@ -1,4 +1,0 @@
-export class UpdateCustomerProfileDto {
-  first_name: string;
-  last_name: string;
-}
