+++ src/microservices/customer/rest/dto/update-customer-profile/index.ts
@@ -1,0 +1,1 @@
+export * from './update-customer-profile.dto';
