+++ src/microservices/customer/rest/dto/update-customer-profile/update-customer-profile.dto.spec.ts
@@ -1,0 +1,11 @@
+import { UpdateCustomerProfileDto } from './update-customer-profile.dto';
+
+describe('UpdateCustomerProfileDto', () => {
+  it('should create a Dto object', () => {
+    const testUpdateCustomerProfileDto = { first_name: 'John', last_name: 'Doe' };
+
+    expect(new UpdateCustomerProfileDto('John', 'Doe')).toEqual(
+        testUpdateCustomerProfileDto,
+    );
+  });
+});
